Squashed commit of the following:
commit 1f50cfd1a1774322d18847b6be5fdac174b06fce
Author: periakteon <masumgokyuz@gmail.com>
Date:   Fri Jun 2 22:56:53 2023 +0300

    conflict resolved

commit cdebe0eee5c5333468a9a2550f2c1009457d3214
Author: Ege Aydemir <egeaydemirr97@gmail.com>
Date:   Fri Jun 2 17:02:29 2023 +0300

    middleware update

commit 20042752e14533692d3273047ccb807e6f9d9b00
Author: Ege Aydemir <egeaydemirr97@gmail.com>
Date:   Fri Jun 2 16:37:53 2023 +0300

    update index
